Refactor decorator-data
